# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Formora API
  version: 0.7.0
  description: HTTP API for the Formora platform.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      summary: Register a user and create a session
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              displayName: Ada Lovelace
              email: ada@example.com
              password: a-secure-password
      responses:
        "201":
          description: The user and session were created.
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: The email address is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Authenticate and create a session
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication succeeded.
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session
      tags: [Authentication]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The current session was revoked or no session existed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
  /api/v1/auth/session:
    get:
      operationId: getSession
      summary: Retrieve the authenticated user
      tags: [Authentication]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The current authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
...
  /api/v1/forms/{formId}/draft:
    parameters:
      - $ref: "#/components/parameters/FormId"
    get:
      operationId: getFormDraft
      summary: Retrieve the authorized form's draft definition
      tags: [Forms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The current draft definition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDraftResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: saveFormDraft
      summary: Replace the authorized form's draft definition
      tags: [Forms]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormDefinition"
      responses:
        "200":
          description: The saved draft definition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDraftResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/forms/{formId}/access-settings:
    parameters:
      - $ref: "#/components/parameters/FormId"
    get:
      operationId: getFormAccessSettings
      summary: Retrieve guest access settings without exposing password hashes
      tags: [Forms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The form's guest access settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormAccessResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateFormAccessSettings
      summary: Configure link or password access for future guests
      tags: [Forms]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormAccessRequest"
      responses:
        "200":
          description: The updated guest access settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormAccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: formora_session
  headers:
...
    SessionCookie:
      description: HTTP-only Formora session cookie.
      schema:
        type: string
  parameters:
...
    FormId:
      name: formId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-fA-F0-9]{24}$"
  responses:
    ValidationError:
      description: The request did not match its boundary schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthenticationRequired:
      description: A valid session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PermissionDenied:
      description: The user does not have the required workspace permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The operation conflicts with the resource's current state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ResponseMeta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
          minLength: 1
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
            requestId:
              type: string
...
    RegisterRequest:
      type: object
      required: [displayName, email, password]
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 80
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthUser:
      type: object
      required: [id, displayName, email, role, createdAt]
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, USER, GUEST]
        createdAt:
          type: string
          format: date-time
    AuthSessionResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [user]
          properties:
            user:
              $ref: "#/components/schemas/AuthUser"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    LogoutResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [success]
          properties:
            success:
              type: boolean
              const: true
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    FormField:
      type: object
      required: [id, type, label, required, placeholder, options]
      properties:
        id:
          type: string
          maxLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
        type:
          type: string
          enum: [text, textarea, email, number, select, checkbox]
        label:
          type: string
          minLength: 1
          maxLength: 120
        required:
          type: boolean
        placeholder:
          type: string
          maxLength: 200
        options:
          type: array
          maxItems: 50
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 120
    FormDefinition:
      type: object
      required: [schemaVersion, fields]
      properties:
        schemaVersion:
          type: integer
          const: 1
        fields:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/FormField"
    FormDraftResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [draft]
          properties:
            draft:
              allOf:
                - $ref: "#/components/schemas/FormDefinition"
                - type: object
                  required: [updatedAt]
                  properties:
                    updatedAt:
                      type: string
                      format: date-time
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    FormAccessRequest:
      type: object
      required: [accessMode]
      properties:
        accessMode:
          type: string
          enum: [LINK, PASSWORD]
        password:
          type: [string, "null"]
          format: password
          minLength: 8
          maxLength: 128
      description: Password is required when accessMode is PASSWORD and ignored for LINK access.
    FormAccessSettings:
      type: object
      required: [accessMode, passwordConfigured]
      properties:
        accessMode:
          type: string
          enum: [LINK, PASSWORD]
        passwordConfigured:
          type: boolean
    FormAccessResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [accessSettings]
          properties:
            accessSettings:
              $ref: "#/components/schemas/FormAccessSettings"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...